refactor(registry): extract integration tests to tests/registry_client.rs
Move client integration tests from client/mod.rs to tests/registry_client.rs.
Keeps client/mod.rs production-only (~285 LOC vs ~1025). Step 0 from SPLIT-PLAN.
