+++ crates/assay-registry/Cargo.toml
@@ -50,6 +50,7 @@ dirs = "6"
 rand = "0.8"
 
 [dev-dependencies]
+base64 = "0.22"
 tempfile.workspace = true
 tokio = { workspace = true, features = ["rt-multi-thread", "macros"] }
 wiremock = "0.6"
